Began the basics of the domain classes
Currently broken.

Added the Item class and its children Node, and Leaf.
Added a DbContext to handle ORM for these classes.
Added a unit test project with a dummy test.
Added DTOs for the domain classes.
Added a controller for API access to the domain classes.

Still to do:
Implement the REST methods properly.
Implement a convenient way to switch between DTOs and domain objects.