Post item should now work
Post needed the adding of a converter to convert JSON objects posted to
either DtoNode or DtoLeaf. DotNet magic allows this to happen without me
changing the actual controller class. The objects sent from the client now
need a a type field to discriminate, but that can be easily changed.
Help from the Salt Monster Frano.